requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
frozenlist==1.5.0
granian==1.6.3
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.6
langchain-community==0.3.4
langchain-core==0.3.15
langchain-text-splitters==0.3.1
langsmith==0.1.139
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.6
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.16
pyyaml==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.37.2
tenacity==9.0.0
typer==0.12.5
typing-extensions==4.12.2
typing-inspect==0.9.0
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.5
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.1
